refactor(gamecreate): move game create to app layer
